Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/mediacapture-record/MediaRecorder-bitrate.https-expected.txt


PASS Passing no bitrate config results in defaults
FAIL Passing bitsPerSecond:0 results in targets close to 0 assert_approx_equals: expected 0 +/- 100000 but got 10192000
PASS Passing only audioBitsPerSecond:0 results in 0 for audio, default for video
PASS Passing only videoBitsPerSecond:0 results in 0 for video, default for audio
FAIL Passing bitsPerSecond:0 overrides audio/video-specific values assert_approx_equals: expected 0 +/- 100000 but got 1100000
FAIL Passing bitsPerSecond overrides audio/video zero values assert_not_equals: got disallowed value 0
FAIL Passing bitsPerSecond sets audio/video bitrate values assert_approx_equals: expected 2000000 +/- 100000 but got 10192000
PASS Passing only audioBitsPerSecond results in default for video
PASS Passing only videoBitsPerSecond results in default for audio
PASS Passing videoBitsPerSecond for audio-only stream still results in something for video
PASS Passing audioBitsPerSecond for video-only stream still results in something for audio
PASS Passing bitsPerSecond for audio-only stream still results in something for video
PASS Passing bitsPerSecond for video-only stream still results in something for audio
PASS Selected default track bitrates are not changed by start()
PASS Passed-in track bitrates are not changed by start()
PASS Passing bitsPerSecond for audio/video stream does not change track bitrates in start()
FAIL Passing bitsPerSecond for audio stream sets video track bitrate to 0 in start() assert_approx_equals: expected 500000 +/- 100000 but got 192000
FAIL Passing bitsPerSecond for video stream sets audio track bitrate to 0 in start() assert_equals: expected 0 but got 192000
